backend_new/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.26.0

# OpenAI
openai==1.6.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Task Queue
celery==5.3.6

# Google Cloud
google-cloud-documentai==2.20.1
google-cloud-storage==2.13.0

# Vector Database
pinecone-client==2.2.4

# Utils
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.0.1
tiktoken==0.5.2
tenacity==8.2.3
aiofiles==23.2.1

# Document Processing
PyMuPDF==1.23.8
pandas==2.1.4
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.1.0
